+++ src/control/mod.rs
@@ -1,3 +1,9 @@
 pub use self::character_controller::RawKinematicCharacterController;
 
+#[cfg(feature = "dim3")]
+pub use self::ray_cast_vehicle_controller::RawDynamicRayCastVehicleController;
+
 mod character_controller;
+
+#[cfg(feature = "dim3")]
+mod ray_cast_vehicle_controller;